fix: Unblock window signals of previous grab if grab is still active
On grabbing a window, signals attached to that window were blocked. If another window was grabbed during this period, the signals were not being unblocked. This caused workspace movements to be ignored. This will now check if a previous grab operation was set, and then attempt to unblock signals on it before proceeding to block signals on the newly-grabbed window.