improv: Add the System aspect of the ECS
We are using GLib as the executor for events. When events are registered, a `GLib.idle_add()` signal is created to execute all registered events. The idle signal will be called repeatedly until no more events remain. This will therefore replace all the random `GLib.idle_add()` calls with a `ext.register_fn()` call that registers a function for execution on our executor.

Window events are now handled through this system. When a window event is being processed, the system shall verify that the window's actor still exists, and handle the window based on the event triggered.