fix: Ignore decoration of Firefox windows via WM_CLASS value
Issue:

The window name is unreliable since Firefox windows come in many different names. This caused certain brands of Firefox to show title bars when it is customized to show the close button inside of the window.

New Behavior:

Now watching for Firefox windows by the WM_CLASS string. Double title bars shouldn't be happening with this change.