fix(autotiler): cleanup applications in dash-to-dock
It looks like the workspace change by id (inside the `auto-tiler->arrange()``) affects closing of windows simultaneously.
Not sure what it is for, but multi window on separate workspaces seems
to close just fine after this change.

Since typescript doesn't allow unused vars. Added an empty if statement
to use the var `workspace` so `ts` will not complain.