improv(packaging): Install schemas to /usr/share/glib-2.0/schemas
This means `dconf-editor` can show descriptions for keys, and Gnome
Control Center won't need a patch to find the schema at a different
path.

Packages like `gnome-shell-extension-desktop-icons-ng` and
`gnome-shell-extension-dashtodock` install their schemas here, so this
seems like a reasonable thing to do.

The changes here are based on the `debian/` directory used for
Ubuntu's `gnome-shell-extension-dashtodock` package.

libglib2.0-0 seems to have a trigger to recompile glib schemas when a file
is changed here, so we don't explicitly handle that.

BREAKING CHANGE: Breaks gnome-control-center, unless the patch to add
the additional schema path is removed.