chore(build): Check for gnome-extensions in Makefile depcheck
`gnome-extensions` is a separate package from `gnome-shell` on some systems (e.g. openSUSE), so when running operations that require it, check if it's present and show the package name if not. If someone doesn't want to install `gnome-extensions`, it's still possible to compile, install, and configure manually using those individual operations.