fix: allow floating exempt windows to be toggled
This allows floating exempt windows to be toggled into tiling mode with
Super + G.  To ensure exempt windows are treated normally for movement
and placement, we must consider windows tagged with ForceTile to be
"is_tilable".

We handle these floating exempt windows separately from other windows in
toggle_floating by checking if they are exempt and then adding the
ForceTile tag to toggle floating on and then removing it when we toggle
it off.

Fixes #1222.