Uncomment import of log utility in utils
Until now when installing pop-os shell into gnome-shell 3.36.9 provided
by canonical the launcher would crash with the error `log.error is not
defined` in journalctl.

The error seems to be that the import of the logging utility (`log.js`)
was never commented in `src/utils.js`. Doing so will solve that problem,
and the launcher is able to start up.