fix: On Gnome 42, close launcher on click outside of it
Fixes https://github.com/pop-os/shell/issues/1386.

This breaks the behavior on older Gnome Shell versions. If we want this
branch to work there as well, it would be easy to make it conditional.

I don't know if there are any other undesired behavior changes here or
the other places `shellReactive: true` is used, but it seems to
basically work as expected now.