move gjs from DEFAULT_FLOAT_RULES to SKIPTASKBAR_EXCEPTIONS
When pop-shell is installed alongside desktop-icons NG and "Show Minimize to Tray Windows" is enabled, gjs shows up as an application in the alt-tab menu (like in this previous bug: https://github.com/pop-os/shell/issues/1056). By completely moving it to the SKIPTASKBAR_EXCEPTIONS it gets resolved/no longer shows up when the option is enabled.